vmaps/requirements.txt
beautifulsoup4>=4.14.2
blinker>=1.9.0
bs4>=0.0.2
cairocffi>=1.7.1
CairoSVG>=2.8.2
certifi>=2025.10.5
cffi>=2.0.0
charset-normalizer>=3.4.4
click>=8.3.0
colorama>=0.4.6
cssselect2>=0.8.0
defusedxml>=0.7.1
Flask>=3.1.2
freetype-py>=2.5.1
idna>=3.11
itsdangerous>=2.2.0
Jinja2>=3.1.6
lxml>=6.0.2
MarkupSafe>=3.0.3
networkx>=3.5
pillow>=12.0.0
pycairo>=1.28.0
pycparser>=2.23
reportlab>=4.4.4
requests>=2.32.5
rlPyCairo>=0.4.0
soupsieve>=2.8
svglib>=1.6.0
tinycss2>=1.4.0
typing_extensions>=4.15.0
urllib3>=2.5.0
webencodings>=0.5.1
Werkzeug>=3.1.3
